Show create review button on spot horizontal list
